requirements.txt
pyserial==3.5
sqlite3
